requirements.txt
aiofiles==23.2.1
aiogram==3.4.1
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
colorama==0.4.6
confluent-kafka==2.3.0
fastapi==0.115.12
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
kafka-python==2.1.4
magic-filter==1.0.12
multidict==6.2.0
packaging==24.2
pluggy==1.5.0
propcache==0.3.1
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.9.0
pytest==8.3.5
pytest-asyncio==0.26.0
python-dotenv==1.1.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uvicorn==0.34.0
yarl==1.18.3
